refactor(backlog): Move final stats logging to Tracker
