+++ backlog/src/BacklogParserWorker.cs
@@ -109,7 +109,7 @@ internal class BacklogParserWorker(
             }
         }
 
-        LogFinalStatistics(alreadyDoneLines, successfulNewLines, failedToProcessLines, csvParseErrors,
+        tracker.LogFinalStatistics(alreadyDoneLines, successfulNewLines, failedToProcessLines, csvParseErrors,
             skippedCsvLineIdentifiers, numAllLinesInCsv);
 
         if (failedToProcessLines.Count > 0 || csvParseErrors.Count > 0)
@@ -118,89 +118,6 @@ internal class BacklogParserWorker(
         }
 
         return 0;
-    }
-
-    private void LogFinalStatistics(List<CsvLine> alreadyDoneLines, List<CsvLine> successfulNewLines,
-        List<(CsvLine line, Exception exception)> failedToProcessLines, List<string> csvParseErrors,
-        List<string> skippedCsvLineIdentifiers, int numAllLinesInCsv)
-    {
-        var numSkippedCsvLines = skippedCsvLineIdentifiers.Count;
-        var markedAsSkipIds = numSkippedCsvLines > 0
-            ? StringJoinFirstFive(skippedCsvLineIdentifiers, ", ")
-            : string.Empty;
-        var successfulFileExtensionBreakdown = string.Join(", ", successfulNewLines.GroupBy(l => l.Extension).Select(g => $"{g.Count()} {g.Key}"));
-        
-        logger.LogInformation("""
-                              ---------------------------
-                              Successfully processed {SuccessfulLinesCount} of {CsvLinesCount} csv lines, of which:
-                                - {NewLinesCount} lines were new ({SuccessfulFileExtensionBreakdown})
-                                - {MarkedToSkipLineCount} lines were marked in the csv to skip ({MarkedToSkipIds})
-                                - {AlreadyDoneLineCount} lines were skipped because they had been processed in a previous run
-                              """,
-            numSkippedCsvLines + alreadyDoneLines.Count + successfulNewLines.Count,
-            numAllLinesInCsv,
-            successfulNewLines.Count,
-            successfulFileExtensionBreakdown,
-            numSkippedCsvLines, markedAsSkipIds,
-            alreadyDoneLines.Count
-        );
-
-        if (csvParseErrors.Count > 0)
-        {
-            logger.LogError("""
-                            ---------------------------
-                            Failed to read {FailedLineCount} lines from the csv:
-                            {FailedLineDetails}
-                            """,
-                csvParseErrors.Count,
-                string.Join(Environment.NewLine, csvParseErrors.Select(error => $"  - {error}"))
-            );
-        }
-
-        if (failedToProcessLines.Count > 0)
-        {
-            var failedIdsGroupedByErrorMessage = failedToProcessLines
-                .GroupBy(f =>
-                    {
-                        return f.exception.Message switch
-                        {
-                            _ when f.exception.Message.StartsWith("Could not find file") => "Could not find file",
-                            _ when f.exception.Message.StartsWith("Couldn't find file with UUID") =>
-                                "Couldn't find file with UUID",
-                            _ when f.exception.Message.EndsWith("was not recognized as a valid DateTime.") =>
-                                "String was not recognized as a valid DateTime",
-                            _ => f.exception.Message
-                        };
-                    },
-                    f => f.line.id);
-
-            var groupedErrorDescriptions = failedIdsGroupedByErrorMessage.Select(groupOfErrors =>
-                $"  - {groupOfErrors.Count()} lines failed with exception message \"{groupOfErrors.Key}\". Ids affected were: ({StringJoinFirstFive(groupOfErrors, ", ")})");
-            var failedFileExtensionBreakdown = string.Join(", ", failedToProcessLines.GroupBy(l => l.line.Extension).Select(g => $"{g.Count()} {g.Key}"));
-
-            logger.LogError("""
-                            ---------------------------
-                            Failed to process {FailedLineCount} lines ({FailedFileExtensionBreakdown}), of which:
-                            {GroupedErrorDescriptions}
-                            """,
-                failedToProcessLines.Count,
-                failedFileExtensionBreakdown,
-                StringJoinFirstFive(groupedErrorDescriptions, Environment.NewLine)
-            );
-        }
-
-        if (failedToProcessLines.Count == 0 && csvParseErrors.Count == 0)
-        {
-            logger.LogInformation("No failed lines");
-        }
-    }
-
-    private static string StringJoinFirstFive(IEnumerable<string> unenumeratedCollection, string separator)
-    {
-        var array = unenumeratedCollection as string[] ?? unenumeratedCollection.ToArray();
-        return array.Length <= 5
-            ? string.Join(separator, array)
-            : string.Join(separator, array.Take(5)) + "...";
     }
 
     private Api.Response CreateResponse(CsvLine csvLine, string mimeType, byte[] sourceContent, bool isStub)
